+++ src/backend/trading-engine/src/main/java/com/alphintra/trading_engine/OrderResponse.java
@@ -7,8 +7,8 @@ import java.util.UUID;
 public class OrderResponse {
     private Long orderId;
     private UUID orderUuid;
-    private Integer userId;
-    private Integer accountId;
+    private Long userId;
+    private Long accountId;
     private String symbol;
     private String side;
     private String orderType;
@@ -32,10 +32,10 @@ public class OrderResponse {
     public void setOrderId(Long orderId) { this.orderId = orderId; }
     public UUID getOrderUuid() { return orderUuid; }
     public void setOrderUuid(UUID orderUuid) { this.orderUuid = orderUuid; }
-    public Integer getUserId() { return userId; }
-    public void setUserId(Integer userId) { this.userId = userId; }
-    public Integer getAccountId() { return accountId; }
-    public void setAccountId(Integer accountId) { this.accountId = accountId; }
+    public Long getUserId() { return userId; }
+    public void setUserId(Long userId) { this.userId = userId; }
+    public Long getAccountId() { return accountId; }
+    public void setAccountId(Long accountId) { this.accountId = accountId; }
     public String getSymbol() { return symbol; }
     public void setSymbol(String symbol) { this.symbol = symbol; }
     public String getSide() { return side; }
