+++ src/backend/trading-engine/src/main/java/com/alphintra/trading_engine/TradingAccountRepository.java
@@ -2,5 +2,5 @@ package com.alphintra.trading_engine;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TradingAccountRepository extends JpaRepository<TradingAccount, Integer> {
+public interface TradingAccountRepository extends JpaRepository<TradingAccount, Long> {
 }
